maybe?

+++ src/python3/pybind_def.h
@@ -17,11 +17,11 @@ template<typename T>
 using py_array_t = pybind11::array_t<T, pybind11::array::c_style | pybind11::array::forcecast>;
 
 
-using pyarray_particles_t = std::tuple<py_array_t<int32_t>, py_array_t<float>, py_array_t<double>,
+using pyarray_particles_t = std::tuple<py_array_t<int32_t>, py_array_t<double>, py_array_t<double>,
                                        py_array_t<double>, py_array_t<double>>;
 
 using pyarray_particles_crt
-    = std::tuple<py_array_t<int32_t> const&, py_array_t<float> const&, py_array_t<double> const&,
+    = std::tuple<py_array_t<int32_t> const&, py_array_t<double> const&, py_array_t<double> const&,
                  py_array_t<double> const&, py_array_t<double> const&>;
 
 template<typename PyArrayInfo>
